add new games
update rainting game
add dates game
