+++ tmdb-app/src/styles/games/game-model-1.css
@@ -1,0 +1,4 @@
+img {
+  width: 60%;
+  height: auto;
+}
