+++ tmdb-app/src/styles/games/game-model-2.css
@@ -1,0 +1,14 @@
+img {
+  /* width: 45%;
+  height: auto; */
+  width: 400px;
+  height: 400px;
+}
+.img-movie {
+  width: 400px;
+  height: 400px;
+}
+
+#hero p {
+  margin-bottom: 5px;
+}
