+++ tmdb-app/src/styles/games/game-model-3.css
@@ -1,0 +1,4 @@
+img {
+  width: 45%;
+  height: auto;
+}
