+++ tmdb-app/src/styles/games/game-model-4.css
@@ -1,0 +1,8 @@
+img {
+  width: 45%;
+  height: auto;
+}
+
+#hero p {
+  margin-bottom: 40px;
+}
