+++ tmdb-app/src/styles/global/main.css
@@ -1,0 +1,47 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  font-family: sans-serif;
+  background-color: black;
+}
+
+a {
+  text-decoration: none;
+  color: white;
+  font-size: 20px;
+}
+
+h1,
+h2,
+h3,
+h4 {
+  font-family: Playfair Display;
+}
+
+h1 {
+  font-size: 50px;
+}
+
+h2 {
+  font-size: 40px;
+}
+
+h3 {
+  font-size: 30px;
+}
+
+h4 {
+  font-size: 24px;
+}
+p {
+  font-size: 20px;
+  font-family: Lora;
+}
+
+img {
+  border-radius: 10px;
+}
