+++ tmdb-app/src/styles/rules.css
@@ -1,0 +1,99 @@
+#hero {
+  /* border: 3px solid violet; */
+  width: 100%;
+  height: 100vh;
+  /* background-image: url("../images/hero.jpg"); /* Replace with your image 
+    background-size: cover; */
+  /* background-position: center; */
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  color: white;
+  text-align: center;
+  padding-top: 80px;
+}
+
+#hero h1 {
+  margin-bottom: 20px;
+}
+
+#hero p {
+  margin-bottom: 40px;
+}
+
+main {
+  color: white;
+  background-color: rgb(35, 35, 35);
+  padding: 20px;
+}
+
+main h2 {
+  text-align: center;
+  margin: 15px;
+}
+
+main p {
+  margin: 10px 0px;
+  text-align: justify;
+}
+
+.paragraph {
+  width: 75%;
+}
+
+.section {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 2rem;
+}
+
+.game-section {
+  display: flex;
+  /* flex-wrap: wrap; */
+  gap: 25px;
+  justify-content: center;
+  align-items: center;
+  margin: 20px 0px;
+}
+
+.game {
+  display: flex;
+  align-items: center;
+
+  border-radius: 20px;
+
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 15px;
+  text-align: center;
+  /* width: 15rem; */
+  /* width: calc(100% / 5); */
+  width: 300px;
+  aspect-ratio: 3 / 4;
+  /* background: #f5f5f5; */
+  position: relative;
+  padding: 1.8rem;
+  border: 5px solid #494949;
+  transition: 0.5s ease-out;
+  overflow: visible;
+}
+
+.description {
+  width: 40%;
+}
+
+.description h3 {
+  text-align: left;
+}
+
+.blue-color {
+  border: 5px solid #007fff;
+}
+
+.colored {
+  color: #007fff;
+}
