--- tmdb-app/src/styles/sevenArtRating.css
@@ -1,181 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: sans-serif;
-  background-color: black;
-}
-
-a {
-  text-decoration: none;
-  color: white;
-}
-
-/* red : e50914 rgb(229, 9, 20)*/
-
-.scrolled {
-  position: fixed;
-  top: 0;
-  width: 100%;
-  padding: 20px 50px;
-  background-color: #333333;
-}
-
-#hero {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  color: white;
-  text-align: center;
-  padding-top: 100px;
-}
-
-#hero h1 {
-  font-size: 48px;
-  margin-bottom: 20px;
-  font-family: Montserrat;
-}
-
-img {
-  width: 45%;
-  height: auto;
-}
-
-#hero p {
-  margin-bottom: 40px;
-}
-
-p {
-  font-size: 24px;
-  font-family: Roboto;
-}
-
-/* *********************** */
-
-.answers {
-  color: white;
-}
-
-main h1 {
-  color: white;
-}
-
-.img {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 15px;
-}
-.rules {
-  background-color: #242424;
-  color: #d3d3d3;
-  cursor: pointer;
-  font-family: Raleway;
-  font-weight: bold;
-  align-self: center;
-}
-
-/************************/
-
-.answers {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 25px;
-  justify-content: center;
-  align-items: center;
-  margin: 20px 0px;
-  align-items: stretch;
-}
-
-.answer {
-  display: flex;
-  align-items: center;
-
-  border-radius: 20px;
-
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 15px;
-  text-align: center;
-
-  width: 40%;
-  position: relative;
-  padding: 1.8rem;
-  border: 5px solid #494949;
-  transition: 0.5s ease-out;
-  overflow: visible;
-
-  /* background-image: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)),
-        url("../images/test.jpg");
-      background-size: cover;
-      background-repeat: no-repeat;
-      background-position: center; */
-}
-
-.white:hover {
-  border-color: white;
-  box-shadow: 0 4px 18px 0 white;
-}
-
-.blue {
-  border: 4px solid #007fff;
-}
-
-.red {
-  border: 4px solid #cc0000;
-}
-
-.green {
-  border: 4px solid #00a056;
-}
-
-/* Couleur #3c7572 */
-
-.game-data {
-  padding: 0 20px;
-  /* border: 3px solid aqua; */
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-  margin: 10px;
-  font-size: 18px;
-}
-
-.game-data * {
-  padding: 10px 20px;
-  font-size: 18px;
-  border: none;
-  border-radius: 5px;
-  text-decoration: none;
-
-  color: white;
-  font-family: Raleway;
-  font-weight: bold;
-}
-
-.mode {
-  background-color: #242424;
-}
-
-.score {
-  background-color: #0362c1;
-}
-
-.validation {
-  background-color: rgb(229, 9, 20);
-}
-
-.questions {
-  background-color: #00a056;
-}
-
-.img-movie {
-  width: 400px;
-  height: 400px;
-}
